fixed bigint stuff

+++ src/main/resources/db/migration/procedures/brs/dashboards/setter/R__get_setter_office_ranking.sql
@@ -42,7 +42,7 @@ BEGIN
                          else ((pd2.project_created_date at time zone 'UTC') at time zone 'US/Mountain') :: date >= up2.start_date
                          end
                                                                                                                                       --if yesterday then we dont include today
-                          and ((ppse2.date_created at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days else (now() at time zone 'US/Mountain')::date end)
+                          and ((ppse2.date_created at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days::integer and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days::integer else (now() at time zone 'US/Mountain')::date end)
 
 --                      and ((pd2.cancelled_date is null) or (pd2.cancelled_date is not null and pd2.cancelled_date > (now() at time zone 'US/Mountain')::date))
                      and o2.id = o.id
@@ -64,14 +64,14 @@ BEGIN
                     then ((pd.project_created_date at time zone 'UTC') at time zone 'US/Mountain') :: date between up.start_date and up.end_date
                     else ((pd.project_created_date at time zone 'UTC') at time zone 'US/Mountain') :: date >= up.start_date
                     end
-                and ((ppse.date_created at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days else (now() at time zone 'US/Mountain')::date end)
+                and ((ppse.date_created at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days::integer and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days::integer else (now() at time zone 'US/Mountain')::date end)
                 and (((case when pd.first_appointment_pitched is not null
                                 then pd.first_appointment_pitched
                             when pd.first_appointment_pitched is null
                                 and pd.first_appointment_missed is not null
                                 then pd.first_appointment_missed
                             else pd.closer_appointment_start
-                               end) at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days else (now() at time zone 'US/Mountain')::date end)
+                               end) at time zone 'UTC') at time zone 'US/Mountain') :: date between ((now() at time zone 'US/Mountain')::date) - p_days::integer and (case when p_time_interval = 'Yesterday' then ((now() at time zone 'US/Mountain')::date) - p_days::integer else (now() at time zone 'US/Mountain')::date end)
                 and (case when pd.first_appointment_pitched is not null
                               then pd.first_appointment_pitched_id in (2,3,1139,1140) --(Pitched, Missed, Pitched - Proposal Not Shown, Pitched - Proposal Shown)
                           when pd.first_appointment_pitched is null
